GIT 提交信息：
功能：精煉認證流程並改善 UI 組件

- 在 `App.vue` 中精煉檢查非空令牌的邏輯，明確檢查空字符串、`null` 和 `undefined`
- 將 `checkUser` 和 `authCheck` 合併為一個函數，明確認證檢查與授權邏輯
- 用 URLSearchParams 替代 `vue-router` 的 useRoute 功能來提取 URL 查詢參數
- 在 `ComponentFooter.vue` 中實作 `showModal` 方法和聯絡表單模態組件
- 更新頁腳樣式以保持一致性並包括聯絡模態組件
- 調整 `ContactFormModal.vue` 中模態的樣式以改善 UI 展示
- 重新排列 `NavbarEl.vue` 項目並改進響應式設計行為
- 修改訊息存儲以在 `message.js` 中返回確認結果
- 在 `PollView.vue` 中實作刪除投票前的條件確認對話框
- 更新 `PollView.vue` 中顯示投票詳情的事件處理
- 重構 `VoteContent.vue` 以響應式處理會員屬性，改善表單互動和錯誤處理
- 從 `VoteContent.vue` 的 Swiper 組件中移除 navigation 屬性

+++ src/views/PollView.vue
@@ -226,12 +226,16 @@ export default {
           }
           break;
         case 'DelPoll':
-          this.handleDeletePoll(id);
           this.message.setMessage({
             title: '刪除投票',
             message: '已成功刪除投票',
           });
-          this.message.showModal(true);
+          (async () => {
+            const result = await this.message.showConfirm();
+            if (result) {
+              this.handleDeletePoll(id);
+            }
+          })();
           break;
         default:
           break;
@@ -313,7 +317,8 @@ export default {
                   :navigation="myPolls.length >= 4 ? navigation : false">
             <swiper-slide v-for="poll in myPolls" :key="poll.id">
               <div
-                   class="relative flex flex-col overflow-hidden bg-white border-2 border-gray-300 rounded-3xl group">
+                  @click="handleDropDown('ShowDetail', poll.id)"
+                   class="relative flex flex-col overflow-hidden bg-white border-2 border-gray-300 rounded-3xl group hover:cursor-pointer">
                 <div class="relative">
                   <div
                        :class="`absolute z-10 ${showCollapse === poll.id ? 'block' : 'hidden'} shadow-lg top-16 right-2 w-44 animate-fade-down animate-once animate-ease-in-out rounded-2xl overflow-hidden`">
